requirements.txt
flask
watchdog
langchain
faiss-cpu  # For CPU usage. If using GPU, change to 'faiss-gpu'
transformers
pdfplumber
python-docx
markdown